backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
sqlalchemy==2.0.30
python-dotenv==1.0.1
pg8000==1.31.2
